Documentation alignment (entity changes).

+++ src/main/java/com/reclamos/backend/repository/SubcategoryRepository.java
@@ -3,5 +3,5 @@ package com.reclamos.backend.repository;
 import com.reclamos.backend.entity.Subcategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubcategoryRepository extends JpaRepository<Subcategory, Integer> {
+public interface SubcategoryRepository extends JpaRepository<Subcategory, Long> {
 }
